+++ src/entry/index.vue
@@ -23,7 +23,7 @@
         @on-dynamic-add='onDynamicAdd'
         @on-dynamic-remove='onDynamicRemove'
       )
-    draggable(
+    vue-drag(
       v-else
       handle='.ep-widget-item-handle'
       draggable='.ep-widget-item'
@@ -50,19 +50,23 @@
         )
 </template>
 <script>
-import { Context, Script, style } from 'epage-core'
-import Epage from 'epage'
-import Draggable from 'vuedraggable'
+import {
+  Context,
+  Script,
+  Event as EpageEvent,
+  style,
+  drag,
+  helper
+} from 'epage-core'
 import EpWidgetItem from './item'
 
-const { helper } = Epage
-const evt = new Epage.Event()
+const evt = new EpageEvent()
 export default {
   on: evt.on.bind(evt),
   off: evt.off.bind(evt),
   components: {
     EpWidgetItem,
-    Draggable
+    VueDrag: drag.VueDrag
   },
 
   data () {
@@ -279,8 +283,6 @@ export default {
     },
 
     onWidgetAdd (schema) {
-      // console.log(323, schema)
-      // this.store.addWidgetChild(schema.key, 0, schema)
       this.store.addWidgetChild(schema.key, schema.children.length, schema.children[0])
       this.$emit('on-add', schema)
     }